Split sidebar table apply input
