+++ Sources/Sidebar/AppKitList/SidebarWorkspaceTableApplyInput.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+/// The latest immutable input delivered by the SwiftUI table bridge.
+@MainActor
+struct SidebarWorkspaceTableApplyInput {
+    let rows: [SidebarWorkspaceTableRowConfiguration]
+    let actions: SidebarWorkspaceTableActions
+    let workspaceIds: [UUID]
+    let selectedWorkspaceId: UUID?
+    let selectedScrollTargetWorkspaceId: UUID?
+}
